Fix incorrect handling of string enconding
Problem: Strings in the CBOR to JSON conversion code were being appended to the
`out` string using `appendFormat` with the `%s` format specifier. This does not
guarantee that UTF8 will be used as the encoding and it will in fact use
whatever encoding is the default on the device that is running the code. On
macOS, for most users, this will in fact be macOS Roman, which leads to things
like `£` characters (anything that's over one byte in UTF8, really) being
appended to the `out` string incorrectly, because they're appended using the
wrong encoding.

Solution: Explicitly use UTF8 encoding in the CBOR to JSON conversion code.