Remove comment about estimating buffer size up front
I'm no longer convinced that this would bring much benefit now that we
at least handle large items more efficiently. If we were to look at
pre-emptive buffer estimation, we should take the time to benchmark and the
effort should be driven by a concrete optimization requirement. This
isn't currently a known hot-spot.